Outline the view_process update steps
It seems like the way the view process steps work is to select all
things that need to have a job run and then run a job. The current
method has a python script that generates and submits some jobs. This
won't really work in nextflow because nextflow will not track the
submitted jobs unless nextflow submits them. This means we have to use a
different approach.

Here we don't use the python script instead we just use nextflow to run
the actual job. I've transfered over most of the key lsf settings,
job group and memory, but haven't set the logging values as nextflow
does that for us. Hopefully that is more or less what is needed.

There are some pretty big gaps in this, first, there is a whole config
thing I need to add, which isn't there yet. Secondly, I haven't tested
anything. And finally there is a mention of a post view step in the
documentation that I don't know what it is yet. I'm also not sure what
parts of the output are needed for those steps. It seems like it is
using the lsf job id and such, but maybe that is to access some output
data?