Improve genome download pipeline
This pipeline seems fixes some issues I observed.

1. This fetches from ENA and NCBI in chunks, meaning one nextflow job
may correspond to many genomes. This is important to ensure we do not
have too many jobs. Too many jobs can cause all sorts of issues with the
cluster/nextflow. Restarting is particularly annoying as you can run
into ulimit issues. Additionally, the cluster can penalize people using
too many jobs.

2. When getting genomes from NCBI it may contain more
scaffolds/patches than uniprot says they used. This is very notable
with the human genome as there are many patches that they do not use.
This can be worked around by filtering the fasta file to contain only
the expected ids.

3. The uniprot proteomes file may specify chromosomes/plasmids that are
not included in the genome downloaded from NCBI. To deal with this the
accession can be downloaded from ENA when filtering the NCBI fasta
files.

To do this I had to change the data that is fetched from the summary
file Uniprot provides. Basically we needed a richer dataset so I created
jsonl files for each source we can fetch from (NCBI and ENA). This
contains enough information to download the correct genomes and the
correct sequences for each genome.

Dealing with ENA is tricky. In some cases when fetching from their API
they will redirect the file to a compressed file on their FTP. This is a
bit annoying as the compression only happens when redirecting. It seems
requests doesn't deal with this easily as the downloaded content will be
the compressed data. To deal with this I download the files with wget
and then inspect them to detect if they are compressed or not. From
there they are parsed and merged. They should be parsed to ensure the
files are valid fasta files and not something else.

I have also added the initial logic to create the actual rfamseq files
based upon the notes in confluence. This still needs testing and tuning
as I haven't tried to touch any memory or retry logic.