Increase amount of fallback to ENA
It seems there are 58 GCA ids that are not present in the NCBI files that
describe their FTP. While I can find some of them if I manually create
the URLs, I cannot find them all. So to deal with that I add a fallback
to ENA where basically this will write out a file that can be used by
download_ena.py to then fetch the sequences from ENA. Hopefully at
least.