Deal with missing NCBI files
It seems there are cases, eg UP000026960/GCA_000182155.2, where the
genomic fasta file does not exist in NCBI. When wget tries to download
them the process fails and produces an empty file. We deal with this by
find all empty files, removing them (so gzip does not break), and then
falling back to ENA based lookups. This seems to be rare, but I haven't
checked. Also I use a hacky shell script to do this instead of a saner
python script, but it works I think.