Deal with WGS sequence issues
It seems some unassembled WGS sequences have the components used
specified as the as GCA accession. This is annoying since no ids in the
genome are that which breaks the assumptions I have in this script. This
deals with that by treating it as using all sequences present in the
file downloaded from NCBI.

This fixes is placed in two files so I don't have to restart the
pipeline from the start and lose all my progress.