Various fixes and improvements
Some fixes from testing on the cluster. Notably, I try to use the
current directory for tmp files. This should be better than the tmp
space as that is more limited and has run out before.

Additionally, this moves to esl-ssplit.pl instead of
esl-randomize-sqfile.pl. It should work the same and be faster.