Look up ids if needed
It is possible to get a GCA id that is not a a complete id (missing the
version). In this case we just assume they are using the latest. This is
likely not to be true, but is reasonable enough as we will be on the
latest that way. There really isn't an easy way to figure out what
version is being used otherwise as well.