Fix version adding
There is one case the GCA id does not have a version and we have to add
it. We should only check if it has a '.' to see if it has a version and
not if it is in the known set of ids. There are cases of valid looking
GCA ids that are not in the NCBI set.