Begin a rewrite
The previous approach made it hard to lookup the metadata, which is
essential to this. I didn't consider that when initially writing this
sadly. This is a partial implementation with incomplete and broken
tests. The point is to have something which can run and try to download
files. It won't work but the errors will be useful.