Improve metadata fetching
This gets closer, but is not yet complete to fetching all needed
metadata. Most of the code is to fetch the description information from
uniprot. We now load and track that since it is what is often used for
description of a genome. This isn't perfect as there are cases when the
description of the genome seems to come from the single chromosome
present and I don't yet deal with that.

This also changes the arguments for building metadata, it now needs the
NCBI assembly info. The code that calls the metadata.build function has
not yet been updated.