Better handle wgs ids
Several issues fixed. First the ids do not always have the same length,
so we generate ones which have the correct length by looking at what the
length of the given id was.

Secondly, we do not treat a wgs record id as a component to find in a
fasta file. This will never happen. Instead what may happen are the
components which are part of that wgs accession may be present. So
instead this will look for those accessions only. Assuming they are
found, we will not have 'extra' ids.