Reduce fetching of WGS sets
Sometimes the scaffolds for a WGS set are in the genome, in this case we
should ignore the individual sequences requested and just use the
requested scaffolds (UP000003336 wants ACTP02000000). Basically if we
see any sequences from a WGS set in a genome we do not fetch the missing
ones. This isn't ideal but should be fine for these error cases.