Add an Accession class
This is meant as a way to represent a standard (non WGS) accession that
is commonly used. This allows for comparing two accessions while
ignoring the versions (which we need to do frequently) as well as
tracking any aliases an accession may have. These are key for selecting
which components of a genome to use.