Rework filter and begin WGS changes
This reworks the fasta_filter module logic so hopefully it gets less
confused about what to download. This requires some yet to be
implemented changes to wgs loading.

In general the logic of this pipeline is to download a GCA/GCF file and
then filter it to see which sequences are used, which are not and which
are not in the file. The missing ones are then downloaded. The previous
version would not detect all sequences were ones that were requested.
This was had at least two causes. First, we ids that were requested did
not have version suffixes (.1, etc) while the ids in the fasta file did.
Additionally, the fasta ids may have used an alias for a requested id.
Secondly, this did not always properly detect that a sequence in the
fasta file was a member of the requested WGS set. It almost handles both
cases properly now.

The remaining work I know of is to fix the WGS logic. Basically we need
to ensure that if the pipeline has been requested to download a WGS set
we have detected at least one sequence from that WGS project. This is
not yet implemented but is doable.