add toponly and merge with curated
Add the `toponly` param to the cmscan step. Update the nf pipeline to fetch the curated text file from the 3d alignments repo, merge this file with the txt file from the pipeline, and then import to the database. Remove step to clan compete rel, as this is unnecessary (as it's already been done) if we add step to use the file for the FTP to import to the rel, pg, fb1 databases.