Fix fetching some genomes
It seems that for some reason this did not actually detect if accessions
matched. Because of this it would not notice that some accessions where
actually requested and did not need to be fetched. This lead to a huge
amount of API and other calls which did not need to happen. By fixing
this it should be much faster to download genomes.