Try to fetch from ENA first
This does several things, first I refactor the download module into
several files. It was too big.

Secondly, we will try to fetch the genome from ENA, if possible, before
trying from NCBI. I've had mixed results in the past with this, but in
some cases it seems to improve things. Hopefully it helps in general. It
seems to more directly fetch the correct genome, or at least the one
with the expected components.

Finally, we have to improve WGS ids parsing. ENA uses ids with .1
version suffixes, even though a WGS id contains an internal version
anyway. This lets the parser work with it and properly strip out the
version, which I think is the right move.