Rework WGS and filtering
This should do a better job at detecting when a sequence is part of a
requested WGS set or not. This involves creating several new classes and
and ways to represent the requested WGS sets. Hopefully this does a
better job at dealing with the requested components. It however does not
have all the previous functionality. I assume it is still needed so
there are some errors left in place.