Improve handling of WGS sets and suppressed genomes
In the case of suppressed genomes this will try to next version of the
genome. This is probably ok. It may still fallback to trying to each out
suppressed component, which isn't great. Ah well.

It now can correctly handle WGS ids vs accessions so it hopefully
downloads WGS sets more efficiently.