Try to improve WGS handling
When we detect if we have found a WGS set in a fasta file this will
allow version mismatches, since sometimes old versions disappear.
However, this did not properly mark the old WGS set as being found. As a
result we would try to fetch extra WGS sets, which we do not want. To
handle this we return both the sequence WGS set and the requested one.
This should prevent duplicate fetching.