Let WGS prefix include trailing 0's
It seems that Uniprot includes this. It is a useful way to distinguish the
different cases, I assume. However, we will parse the ids as soon as
possible and so this should tolerate things with trailing zeros. The
other solution is to parse it as a WgsSequenceId and then check if
the sequence id is 0 and if so then to treat it as prefix instead.
I think this is a bit cleaner.