Rework core download logic
This reworks the core download logic. There were a lot of problems with
the previous approach where it had edge cases I couldn't solve in the
current design. I removed a lot of logic, eg the DownloadMethod bit, and
have hopefully come up with a cleaner approach. This should allow for a
simpler download method. It will allow for batch lookups of missed
sequences and eases up the difference between having a genome accession
and not.

I've tested some cases and it works correctly and a previous edge case
was now simple.