Properly deal with fasta files containing duplicate ids
Seems that ENA does produce fasta files with duplicate ids, at least in
the case of GCA_000757425.2. To deal with this we swallow the duplicate
sequence but log it.

I also added duplicate checking higher up so that in the case something
has gone wrong and we are fetching duplicates this will deal with that.