Handle incorrectly compressed files
It seems that sometimes we get files from ENA that are not actually
compressed. Or they cannot be decompressed at the least. This handles
that more gracefully. Basically we will try to decompress and if that
fails for some reason we will raise a FetchError, this will allow us to
retry with fetching from another resource like NCBI.