Improve fasta filtering
This is now does a better job tracking what was missed and what needs to
be fetched. For example, if the genome contains no unplaced sequences,
which are a member of a requested WGS set it will correctly reduce the
steps to fetch to be the requested WGS set only. This should make
downloading and udpating data much faster.