Rewrite to drastic speed up
Basically I reworked the entire download logic to be much faster. There
are a few changes that make this much faster.

1. Use esl-sfetch to extract sequences. This is much faster to extract
the requested ids from a fasta file. This replaces the logic of parsing
the fasta file in python.

2. Use a dict not a list to store all NcbiSequenceInfo objects. This is
probably the most important change for speed. Basically without this the
process would do N^2 number of checks to see if any given accession was
requested. Using a dict makes this much faster in the case of large
genomes.

The net result is that it is now possible to filter some proteomes, eg
UP000032142, in <1 minute, where before it was >24 hours.

This does mean the pipeline requires seqkit and esl-sfetch to work
though.