Use all sequences for GCA/GCF
As the comment explains this is a major change to how we fetch
sequences. However, this should increase the reliability quite a bit.
Which is the major concern of the current code.

The most concerning effects of this are.

1. This could miss sequences. This can happen if the component is not a
   member of the genome. This is very odd and I haven't seen it. The
   closest I've seen is using things which are more or less assembled,
   which isn't ideal but hopefully this is close enough.
2. This could use the wrong identifiers for a genome. As an example, the
   components could be requesting a WGS set. While those sequences are
   part of the genome, the ids used are not the WGS ids but instead
   different ones. This could be an issue for people mapping from Rfam
   results to UniProt results.

Some other minor effects are.

1. We will have too many sequences. Basically UniProt may exclude some
   partially assembled sequences we use. That isn't really a problem in
   my view.

Overall I think this should be done because it should reduce the failure
modes a lot. Making this reliable is more important than ensure ids are
exactly the same and missing sequences seems unlikely.