Incomplete rewrite
This rewrites much of the logic of the pipeline. This now works very
differently.

First, it uses the JSON api from Uniprot. This is much easier to work
with. It also use models the data from the API and does not create new
custom data structures, for better or worse. Because of this all uniprot
tests are broken currently.

Secondly, this adds wrappers to ENA and NCBI to figure out how to fetch
data for them. The wrappers do not actually fetch data, simply say where
it may be found.

Next this reworks the downloader to use cleaner but different logic.
Given something which is a GCA or GCF it will simply download the entire
genome and not check the components. This could be risky but should be
fine. We can validate if the genome is as expected later.

The fallback logic is changed to be clearer and not use as many
exception handlers, hopefully it should be easier to follow now.

It can do a better job handling outdated genomes, in that it will always
try to use the latest. This is generally safe, I hope.

It also can fetch custom proteomes, not just the reference ones. This is
because we want to add additional viral genomes, as we have poor
coverage there. This adds some logic to the pipeline and the python code
to fetch and deduplicate such entries. We use the data in PIR to select
which additional proteomes to add.

It also does not fail if any one proteome download fails, instead this
is written to a failure file, so they can be inspected later.

This is not yet incomplete in that it likely does not handle WGS sets,
and does not yet produce metadata. However, it is far enough along that
it is worth running on the entire dataset and seeing what happens. The
failure cases can then be inspected. Also, there is a considerable about
of duplicated, unneeded and likely incorrect code that will have to be
cleaned up once the core logic is figured out. This is only being
committed to see what parts need to be kept.