Improve sequence chunking
This should be a much faster way to chunk the sequences. The previous
method used a perl script from elsewhere in Rfam, but I much prefer to
keep everything self-contained and in as few languages as possible.

The new method works by splitting the sequences in a seqstat file into
chunks of roughly the given size. This is done by using the sequence
length in the file as an estimate of it size, that should be safe,
probably. These chunks are then used as input to esl-sfetch to pull out
the given sequences and build the target chunks.