First work on updating text search
The full region export takes way too much memory and way too long
(<100GB of memory and hours of time). This is work to fix that. The
queries can be improved and the pipeline speed up. This starts with
reducing the amount of data fetched, parallelizing the queries and using
caching for values that should just be cached.